Clean up rtw_network_alloc()
Replace driver's self-implemented buffer management and use kmalloc()
to allocate struct wlan_network on demand. This simplifies the buffer
handling and also reduces the memory footprint as the old driver would
allocate almost 128KB (128 entries) for this.

In addition, remove unused counter for number of scanned networks. The
was counted, but nothing was done with the result.

Signed-off-by: Jes Sorensen <Jes.Sorensen@redhat.com>