Extending size of recv_frame does not guarantee alignment
Why oh why???? Setting a minimum size of the struct (which is purely an
internal kernel struct fwiw) does not guarantee alignment to the
minimum size when the struct is carved out of a big block of
memory from a single large allocation.

Remove this pointless entry in union recv_frame, which allows us to
remove the union completely and return to a plain struct (in the
follow-on patch).

Signed-off-by: Jes Sorensen <Jes.Sorensen@redhat.com>