Guard against oops for bad BIT()
One of the GitHub users reported an oops when connected to a particular
network. The traceback led to a BIT() statement with a vaiable argument.
Guarding against that u32 argument having a value greater than 31 fixed
the oops. This patch adds a wrapper around those instances of BIT() that
might get out of bounds.

Signed-off-by: Larry Finger <Larry.Finger@lwfinger.net>